some logic added to payment
